moving server.py structure to estadoserver.py
