+++ rpiZero/modos/local_estadocalibracao.py
